refactor: keep own copy of serializers
